+++ src/main/java/com/deal4u/fourplease/domain/notification/NotificationSender.java
@@ -1,7 +1,7 @@
 package com.deal4u.fourplease.domain.notification;
 
 import com.deal4u.fourplease.domain.notification.email.HtmlEmailMessage;
-import com.deal4u.fourplease.domain.notification.platorm.message.PushNotificationMessage;
+import com.deal4u.fourplease.domain.notification.pushnotification.message.PushNotificationMessage;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.stereotype.Component;
