--- src/main/java/com/deal4u/fourplease/domain/notification/platorm/repository/PushNotificationRepository.java
@@ -1,8 +1,0 @@
-package com.deal4u.fourplease.domain.notification.platorm.repository;
-
-import com.deal4u.fourplease.domain.notification.platorm.entity.PushNotification;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface PushNotificationRepository extends JpaRepository<PushNotification, Long> {
-
-}
