+++ src/main/java/com/deal4u/fourplease/domain/notification/pushnotification/dto/PushNotificationCreateCommand.java
@@ -1,4 +1,4 @@
-package com.deal4u.fourplease.domain.notification.platorm.dto;
+package com.deal4u.fourplease.domain.notification.pushnotification.dto;
 
 public record PushNotificationCreateCommand(Long memberId, String message) {
