+++ src/main/java/com/deal4u/fourplease/domain/notification/pushnotification/entity/PushNotification.java
@@ -1,4 +1,4 @@
-package com.deal4u.fourplease.domain.notification.platorm.entity;
+package com.deal4u.fourplease.domain.notification.pushnotification.entity;
 
 import com.deal4u.fourplease.domain.BaseDateEntity;
 import jakarta.persistence.Entity;
