+++ src/main/java/com/deal4u/fourplease/domain/notification/pushnotification/exception/PushNotificationMappingException.java
@@ -1,4 +1,4 @@
-package com.deal4u.fourplease.domain.notification.platorm.exception;
+package com.deal4u.fourplease.domain.notification.pushnotification.exception;
 
 public class PushNotificationMappingException extends RuntimeException {
