+++ src/main/java/com/deal4u/fourplease/domain/notification/pushnotification/message/PushMessageBody.java
@@ -1,4 +1,4 @@
-package com.deal4u.fourplease.domain.notification.platorm.message;
+package com.deal4u.fourplease.domain.notification.pushnotification.message;
 
 import java.util.Map;
