+++ src/main/java/com/deal4u/fourplease/domain/notification/pushnotification/service/PushNotificationService.java
@@ -1,8 +1,8 @@
-package com.deal4u.fourplease.domain.notification.platorm.service;
+package com.deal4u.fourplease.domain.notification.pushnotification.service;
 
-import com.deal4u.fourplease.domain.notification.platorm.dto.PushNotifications;
-import com.deal4u.fourplease.domain.notification.platorm.mapper.PushNotificationMapper;
-import com.deal4u.fourplease.domain.notification.platorm.message.PushNotificationMessage;
+import com.deal4u.fourplease.domain.notification.pushnotification.dto.PushNotifications;
+import com.deal4u.fourplease.domain.notification.pushnotification.mapper.PushNotificationMapper;
+import com.deal4u.fourplease.domain.notification.pushnotification.message.PushNotificationMessage;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
